Ajuste qualificador de precisao
Ajuste qualificador de precisao que foi declarado como mediump mas no shaders nao estava causando erro no build

+++ examples/ufabcbird/assets/bg.vert
@@ -2,7 +2,7 @@
 
 layout (location = 0) in vec2 inPosition;
 
-uniform float time;
+uniform mediump float time;
 out vec4 fragColor;
 void main() {
     gl_Position = vec4(inPosition, 1.0, 1);
